Prevent deleting other users comment in DestroyTest

+++ app/Http/Controllers/CommentController.php
@@ -68,6 +68,9 @@ class CommentController extends Controller
      */
     public function destroy(Comment $comment)
     {
+        if (auth()->user()->id !== $comment->user->id) {
+            abort(403);
+        }
         $comment->delete();
 
         return to_route('posts.show', $comment->post_id);
